+++ lib/twilio-ruby/rest/taskrouter/v1/workspace/workflow/workflow_real_time_statistics.rb
@@ -135,6 +135,7 @@ module Twilio
                 @properties = {
                     'account_sid' => payload['account_sid'],
                     'longest_task_waiting_age' => payload['longest_task_waiting_age'].to_i,
+                    'longest_task_waiting_sid' => payload['longest_task_waiting_sid'],
                     'tasks_by_priority' => payload['tasks_by_priority'],
                     'tasks_by_status' => payload['tasks_by_status'],
                     'total_tasks' => payload['total_tasks'].to_i,
@@ -173,6 +174,12 @@ module Twilio
               # @return [String] The age of the longest waiting Task
               def longest_task_waiting_age
                 @properties['longest_task_waiting_age']
+              end
+
+              ##
+              # @return [String] The SID of the longest waiting Task
+              def longest_task_waiting_sid
+                @properties['longest_task_waiting_sid']
               end
 
               ##