+++ spec/integration/api/v2010/account/usage/trigger_spec.rb
@@ -143,13 +143,13 @@ describe 'Trigger' do
     expect {
       @client.api.v2010.accounts('ACXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX') \
                        .usage \
-                       .triggers.create(callback_url: 'https://example.com', trigger_value: 'trigger_value', usage_category: 'answering-machine-detection')
+                       .triggers.create(callback_url: 'https://example.com', trigger_value: 'trigger_value', usage_category: 'agent-conference')
     }.to raise_exception(Twilio::REST::TwilioError)
 
     values = {
         'CallbackUrl' => 'https://example.com',
         'TriggerValue' => 'trigger_value',
-        'UsageCategory' => 'answering-machine-detection',
+        'UsageCategory' => 'agent-conference',
     }
     expect(
     @holodeck.has_request?(Holodeck::Request.new(
@@ -186,7 +186,7 @@ describe 'Trigger' do
 
     actual = @client.api.v2010.accounts('ACXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX') \
                               .usage \
-                              .triggers.create(callback_url: 'https://example.com', trigger_value: 'trigger_value', usage_category: 'answering-machine-detection')
+                              .triggers.create(callback_url: 'https://example.com', trigger_value: 'trigger_value', usage_category: 'agent-conference')
 
     expect(actual).to_not eq(nil)
   end