+++ spec/integration/taskrouter/v1/workspace/task_queue/task_queue_real_time_statistics_spec.rb
@@ -32,6 +32,7 @@ describe 'TaskQueueRealTimeStatistics' do
       %q[
       {
           "longest_task_waiting_age": 100,
+          "longest_task_waiting_sid": "WTaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "task_queue_sid": "WQaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "tasks_by_status": {
               "reserved": 0,