+++ spec/integration/taskrouter/v1/workspace/workflow/workflow_real_time_statistics_spec.rb
@@ -32,6 +32,7 @@ describe 'WorkflowRealTimeStatistics' do
       %q[
       {
           "longest_task_waiting_age": 100,
+          "longest_task_waiting_sid": "WTaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "url": "https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Workflows/WWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/RealTimeStatistics",
           "tasks_by_priority": {},
           "workspace_sid": "WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",