+++ spec/integration/taskrouter/v1/workspace/workflow/workflow_statistics_spec.rb
@@ -47,7 +47,7 @@ describe 'WorkflowStatistics' do
           },
           "realtime": {
               "longest_task_waiting_age": 0,
-              "longest_task_waiting_sid": null,
+              "longest_task_waiting_sid": "WTaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
               "tasks_by_status": {
                   "assigned": 1,
                   "pending": 0,